추천 기능 구현

+++ src/main/java/com/springproject/model/Likey.java
@@ -20,7 +20,7 @@ public class Likey {
     private Long userNumber;
 
     @Column
-    private int evaluationId;
+    private Long evaluationId;
 
     @Column(length = 50)
     private String userIp;
